[Add] PointOperations
PointOperations allow us to change the current
points during the game.
Updated Dialogue, Option and GameState to fit
the new PointOperations.
Updated GameDialogue.xml to exemplify the new
PointOperations as well as display a quick test for them.
Updated PointGateTutorial.xml to alert about escaping
the Less Than sign in the expressions.